Remove code deplication
