+++ src/UNL/UCBCN/Manager/CreateEvent.php
@@ -1,39 +1,18 @@
 <?php
 namespace UNL\UCBCN\Manager;
 
+use UNL\UCBCN\Manager\EventForm as EventForm;
 use UNL\UCBCN as BaseUCBCN;
-use UNL\UCBCN\Calendar as CalendarModel;
-use UNL\UCBCN\Calendar\EventTypes;
-use UNL\UCBCN\FileUpload;
-use UNL\UCBCN\Location;
 use UNL\UCBCN\Locations;
 use UNL\UCBCN\Event;
 use UNL\UCBCN\Event\EventType;
 use UNL\UCBCN\Event\Occurrence;
-use UNL\UCBCN\User;
-use UNL\UCBCN\Permission;
 
-class CreateEvent extends PostHandler
+class CreateEvent extends EventForm
 {
-    public $options = array();
-    public $calendar;
-    public $event;
-    public $post;
-
     public function __construct($options = array()) 
     {
-        $this->options = $options + $this->options;
-        $this->calendar = CalendarModel::getByShortName($this->options['calendar_shortname']);
-
-        if ($this->calendar === FALSE) {
-            throw new \Exception("That calendar could not be found.", 404);
-        }
-
-        $user = Auth::getCurrentUser();
-        if (!$user->hasPermission(Permission::EVENT_CREATE_ID, $this->calendar->id)) {
-            throw new \Exception("You do not have permission to create an event on this calendar.", 403);
-        }
-
+        parent::__construct($options);
         $this->event = new Event;
     }
 
@@ -49,23 +28,6 @@ class CreateEvent extends PostHandler
 
         # redirect
         return $this->calendar->getManageURL(TRUE);
-    }
-
-    private function setEventData($post_data, $files) 
-    {
-        $this->event->title = empty($post_data['title']) ? NULL : $post_data['title'];
-        $this->event->subtitle = empty($post_data['subtitle']) ? NULL : $post_data['subtitle'];
-        $this->event->description = empty($post_data['description']) ? NULL : $post_data['description'];
-
-        $this->event->listingcontactname = empty($post_data['contact_name']) ? NULL : $post_data['contact_name'];
-        $this->event->listingcontactphone = empty($post_data['contact_phone']) ? NULL : $post_data['contact_phone'];
-        $this->event->listingcontactemail = empty($post_data['contact_email']) ? NULL : $post_data['contact_email'];
-
-        $this->event->webpageurl = empty($post_data['website']) ? NULL : $post_data['website'];
-        $this->event->approvedforcirculation = isset($post_data['private_public']) && $post_data['private_public'] == 'private' ? 0 : 1;
-
-        # for extraneous data aside from the event (location, type, etc)
-        $this->post = $post_data;
     }
 
     private function validateEventData($post_data, $files) 
@@ -250,6 +212,20 @@ class CreateEvent extends PostHandler
         return $this->event;
     }
 
+    public function getUserLocations()
+    {
+        $user = Auth::getCurrentUser();
+        return new Locations(array('user_id' => $user->uid));
+    }
+
+	public function getStandardLocations($display_order)
+	{
+		return new Locations(array(
+			'standard' => true,
+			'display_order' => $display_order,
+		));
+	}
+
     /**
      * Add a location
      * 
@@ -292,24 +268,5 @@ class CreateEvent extends PostHandler
         $location->insert();
         
         return $location;
-    }
-
-    public function getEventTypes()
-    {
-        return new EventTypes(array());
-    }
-    
-    public function getUserLocations()
-    {
-        $user = Auth::getCurrentUser();
-        return new Locations(array('user_id' => $user->uid));
-    }
-    
-    public function getStandardLocations($display_order)
-    {
-        return new Locations(array(
-            'standard' => true,
-            'display_order' => $display_order,
-        ));
     }
 }
