+++ src/UNL/UCBCN/Manager/EditEvent.php
@@ -1,38 +1,19 @@
 <?php
 namespace UNL\UCBCN\Manager;
 
-use UNL\UCBCN\Calendar as CalendarModel;
-use UNL\UCBCN\Calendar\EventTypes;
-use UNL\UCBCN\Location;
-use UNL\UCBCN\Locations;
-use UNL\UCBCN\Permission;
+use UNL\UCBCN\Manager\EventForm as EventForm;
 use UNL\UCBCN\Event;
 use UNL\UCBCN\Event\EventType;
-use UNL\UCBCN\Event\Occurrence;
-use UNL\UCBCN\User;
-use UNL\UCBCN\FileUpload;
+use UNL\UCBCN\Calendar as CalendarModel;
 
-class EditEvent extends PostHandler
+class EditEvent extends EventForm
 {
-    public $options = array();
-    public $calendar;
-    public $event;
-    public $post;
     public $on_main_calendar;
     public $page;
 
     public function __construct($options = array()) 
     {
-        $this->options = $options + $this->options;
-        $this->calendar = CalendarModel::getByShortName($this->options['calendar_shortname']);
-        if ($this->calendar === FALSE) {
-            throw new \Exception("That calendar could not be found.", 404);
-        }
-
-        $user = Auth::getCurrentUser();
-        if (!$user->hasPermission(Permission::EVENT_EDIT_ID, $this->calendar->id)) {
-            throw new \Exception("You do not have permission to edit events on this calendar.", 403);
-        }
+        parent::__construct($options);
 
         $this->event = Event::getByID($this->options['event_id']);
         if ($this->event === FALSE) {
@@ -64,34 +45,6 @@ class EditEvent extends PostHandler
 
         $this->flashNotice(parent::NOTICE_LEVEL_SUCCESS, 'Event Updated', 'The event "' . $this->event->title . '" has been updated.');
         return $this->event->getEditURL($this->calendar);
-    }
-
-    public function getEventTypes()
-    {
-        return new EventTypes(array());
-    }
-
-    public function getLocations()
-    {
-        $user = Auth::getCurrentUser();
-        return new Locations(array('user_id' => $user->uid));
-    }
-
-    private function setEventData($post_data, $files)
-    {
-        $this->event->title = empty($post_data['title']) ? NULL : $post_data['title'];
-        $this->event->subtitle = empty($post_data['subtitle']) ? NULL : $post_data['subtitle'];
-        $this->event->description = empty($post_data['description']) ? NULL : $post_data['description'];
-
-        $this->event->listingcontactname = empty($post_data['contact_name']) ? NULL : $post_data['contact_name'];
-        $this->event->listingcontactphone = empty($post_data['contact_phone']) ? NULL : $post_data['contact_phone'];
-        $this->event->listingcontactemail = empty($post_data['contact_email']) ? NULL : $post_data['contact_email'];
-
-        $this->event->webpageurl = empty($post_data['website']) ? NULL : $post_data['website'];
-        $this->event->approvedforcirculation = $post_data['private_public'] == 'private' ? 0 : 1;
-
-        # for extraneous data aside from the event (location, type, etc)
-        $this->post = $post_data;
     }
 
     private function validateEventData($post_data, $files)
